requirements.txt
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.11.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-sqlalchemy==0.18.0
more-itertools==8.3.0
packaging==20.4
pathlib2==2.3.5
pkg-resources==0.0.0
pluggy==0.13.1
py==1.8.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.2
pytz==2020.1
redis==3.5.2
six==1.14.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.5
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==1.2.0
